Split LibraryPage into filter, controller card, and view modules.
Extract pure library filter matching, ControllerGameCard, and
LibraryControllerView while keeping LibraryPage as the coordinator.
Share appendUnique/appendImageType/gameMatchesActiveSession with HomePage
via controllerCatalogUi.
