+++ opennow-stable/src/renderer/src/components/LibraryPage.tsx
@@ -1,48 +1,30 @@
-import { Library, Search, Clock, Gamepad2, Loader2, ArrowUpDown, MoreHorizontal, Menu, Filter, ChevronDown, X } from "lucide-react";
+import { Library, Search, Clock, Gamepad2, Loader2, ArrowUpDown, Filter, ChevronDown, X } from "lucide-react";
 import { memo, useEffect, useMemo, useRef, useState } from "react";
 import type { JSX } from "react";
 import { AnimatePresence, m } from "motion/react";
 import type { CatalogSortOption, GameInfo } from "@shared/gfn";
-import { getStoreDisplayName, getStoreIconComponent, normalizeStoreKey } from "./GameCard";
 import { GameCardListItem, useCatalogCardActionsRef } from "./GameCardListItem";
 import type { PlaytimeData } from "../lib/gameCatalog";
+import { getControllerFeaturedGames } from "../lib/controllerCatalogUi";
+import {
+  gameMatchesLibraryFilters,
+  gameMatchesStoreFilter,
+  getControllerStoreFilterItems,
+  getLibraryFilterGroups,
+  getLibraryFilterOptionById,
+  type ControllerStoreFilterItem,
+  type LibraryFilterOption,
+} from "../lib/libraryFilters";
 import { useTranslation } from "../i18n";
 import { formatCatalogLastPlayed } from "../utils/lastPlayedFormat";
 import { controllerButton, readControllerGamepadButtons } from "../utils/controllerGamepad";
-import { pageTransition, panelSpring } from "./MotionProvider";
+import { pageTransition } from "./MotionProvider";
 import { SelectDropdown } from "./ui/SelectDropdown";
+import { LibraryControllerView } from "./library/LibraryControllerView";
 
 const CONTROLLER_HERO_ROTATION_MS = 8000;
 const CONTROLLER_MOVE_REPEAT_MS = 140;
 const CONTROLLER_Y_HOLD_MS = 350;
-
-const CONTROLLER_HERO_BACKGROUND_KEYS = [
-  "MARQUEE_HERO_IMAGE",
-  "FEATURE_IMAGE",
-  "HERO_IMAGE",
-  "TV_BANNER",
-  "KEY_ART",
-  "KEY_IMAGE",
-] as const;
-
-interface ControllerStoreFilterItem {
-  id: string;
-  title: string;
-}
-
-interface LibraryFilterOption {
-  id: string;
-  label: string;
-  count: number;
-}
-
-interface LibraryFilterGroup {
-  id: string;
-  label: string;
-  options: LibraryFilterOption[];
-}
-
-type LibraryTranslation = (key: string, values?: Record<string, string | number | boolean | null | undefined>) => string;
 
 export interface LibraryPageProps {
   games: GameInfo[];
@@ -66,346 +48,6 @@ export interface LibraryPageProps {
   activeSessionAppIds?: number[];
   onPreviousControllerPage?: () => void;
   onNextControllerPage?: () => void;
-}
-
-function appendUnique(values: string[], candidate: string | undefined): void {
-  if (!candidate || values.includes(candidate)) return;
-  values.push(candidate);
-}
-
-function appendImageType(values: string[], game: GameInfo, type: string): void {
-  for (const candidate of game.imageUrlsByType?.[type] ?? []) {
-    appendUnique(values, candidate);
-  }
-}
-
-function getControllerHeroBackgroundCandidates(game: GameInfo): string[] {
-  const candidates: string[] = [];
-  for (const type of CONTROLLER_HERO_BACKGROUND_KEYS) {
-    appendImageType(candidates, game, type);
-  }
-  appendUnique(candidates, game.heroImageUrl);
-  appendUnique(candidates, game.imageUrl);
-  for (const candidate of game.screenshotUrls ?? []) appendUnique(candidates, candidate);
-  appendUnique(candidates, game.screenshotUrl);
-  return candidates;
-}
-
-function getControllerHeroLogoUrl(game: GameInfo): string | undefined {
-  return game.imageUrlsByType?.GAME_LOGO?.find(Boolean);
-}
-
-function getGameLogoUrl(game: GameInfo): string | undefined {
-  return game.imageUrlsByType?.GAME_LOGO?.find(Boolean);
-}
-
-function getControllerFeaturedGames(featuredGames: GameInfo[], fallbackGames: GameInfo[]): GameInfo[] {
-  const source = featuredGames.length > 0 ? featuredGames : fallbackGames;
-  return source.slice(0, 6);
-}
-
-function getGameStoreSummary(game: GameInfo, fallback: string): string {
-  const stores = [...new Set((game.availableStores?.length ? game.availableStores : game.variants.map((variant) => variant.store)).filter(Boolean))];
-  if (stores.length === 0) return fallback;
-  const visible = stores.slice(0, 3).join(", ");
-  return stores.length > 3 ? `${visible} +${stores.length - 3}` : visible;
-}
-
-function getSelectedVariantStoreLabel(game: GameInfo, selectedVariantId: string | undefined, fallback: string): string {
-  const selectedVariant = game.variants.find((variant) => variant.id === selectedVariantId)
-    ?? game.variants[game.selectedVariantIndex]
-    ?? game.variants[0];
-  return selectedVariant?.store ? getStoreDisplayName(selectedVariant.store) : fallback;
-}
-
-function getPlayerSummary(game: GameInfo): string | null {
-  const parts: string[] = [];
-  if (game.maxLocalPlayers && game.maxLocalPlayers > 0) parts.push(`Local ${game.maxLocalPlayers}`);
-  if (game.maxOnlinePlayers && game.maxOnlinePlayers > 0) parts.push(`Online ${game.maxOnlinePlayers}`);
-  return parts.length > 0 ? parts.join(" / ") : null;
-}
-
-function gameMatchesActiveSession(game: GameInfo, activeSessionAppIds: number[]): boolean {
-  if (activeSessionAppIds.length === 0) return false;
-  const appIds = new Set(activeSessionAppIds.map(String));
-  if (game.launchAppId && appIds.has(game.launchAppId)) return true;
-  if (appIds.has(game.id)) return true;
-  return game.variants.some((variant) => appIds.has(variant.id));
-}
-
-function gameMatchesStoreFilter(game: GameInfo, filterId: string): boolean {
-  if (filterId === "library") return true;
-  const store = filterId.slice("store:".length);
-  return game.variants.some((variant) => variant.store === store) || (game.availableStores ?? []).includes(store);
-}
-
-function normalizeLibraryFilterValue(value: string | undefined): string {
-  return (value ?? "")
-    .trim()
-    .toLowerCase()
-    .replace(/['’]/g, "")
-    .replace(/[^a-z0-9]+/g, "-")
-    .replace(/^-+|-+$/g, "");
-}
-
-function formatLibraryFilterLabel(value: string): string {
-  return value
-    .trim()
-    .replace(/[_-]+/g, " ")
-    .replace(/\s+/g, " ")
-    .toLowerCase()
-    .replace(/\b\w/g, (match) => match.toUpperCase());
-}
-
-function getGameCatalogSkuText(game: GameInfo): string {
-  const values = Object.values(game.catalogSkuStrings ?? {});
-  const parts: string[] = [];
-  for (const value of values) {
-    if (typeof value === "string") {
-      parts.push(value);
-    } else if (Array.isArray(value)) {
-      parts.push(...value.filter((item): item is string => typeof item === "string"));
-    }
-  }
-  return parts.join(" ");
-}
-
-function gameRequiresInstallToPlay(game: GameInfo): boolean {
-  const haystack = [
-    game.playType,
-    game.playabilityState,
-    ...(game.featureLabels ?? []),
-    getGameCatalogSkuText(game),
-  ].join(" ").toLowerCase();
-  return haystack.includes("install");
-}
-
-function getGamePlayTypeFilter(game: GameInfo, t: LibraryTranslation): LibraryFilterOption | null {
-  if (gameRequiresInstallToPlay(game)) {
-    return { id: "play:install-to-play", label: t("library.filterOptions.installToPlay"), count: 0 };
-  }
-  if (!game.playType?.trim()) return null;
-  const normalized = normalizeLibraryFilterValue(game.playType);
-  if (!normalized) return null;
-  if (normalized.includes("instant") || normalized.includes("stream") || normalized.includes("cloud")) {
-    return { id: "play:cloud-launch", label: t("library.filterOptions.cloudLaunch"), count: 0 };
-  }
-  const label = formatLibraryFilterLabel(game.playType);
-  const key = normalizeLibraryFilterValue(label);
-  return key ? { id: `play:${key}`, label, count: 0 } : null;
-}
-
-function getGameStoreFilters(game: GameInfo): Array<{ id: string; label: string }> {
-  const stores = game.availableStores?.length ? game.availableStores : game.variants.map((variant) => variant.store);
-  const seen = new Set<string>();
-  const filters: Array<{ id: string; label: string }> = [];
-  for (const store of stores) {
-    if (!store.trim()) continue;
-    const key = normalizeStoreKey(store);
-    if (seen.has(key)) continue;
-    seen.add(key);
-    filters.push({ id: `platform:${key}`, label: getStoreDisplayName(store) });
-  }
-  return filters;
-}
-
-function getControlFilter(control: string, t: LibraryTranslation): { id: string; label: string } | null {
-  const normalized = normalizeLibraryFilterValue(control);
-  if (!normalized) return null;
-  if (normalized.includes("keyboard") || normalized.includes("mouse")) {
-    return { id: "controls:keyboard-mouse", label: t("library.filterOptions.keyboardMouse") };
-  }
-  if (normalized.includes("gamepad") || normalized.includes("controller")) {
-    return { id: "controls:controller", label: t("library.filterOptions.controller") };
-  }
-  if (normalized.includes("touch")) {
-    return { id: "controls:touch", label: t("library.filterOptions.touch") };
-  }
-  const label = formatLibraryFilterLabel(control);
-  const key = normalizeLibraryFilterValue(label);
-  return key ? { id: `controls:${key}`, label } : null;
-}
-
-function getGameControlFilters(game: GameInfo, t: LibraryTranslation): Array<{ id: string; label: string }> {
-  const controls = [
-    ...(game.supportedControls ?? []),
-    ...game.variants.flatMap((variant) => variant.supportedControls),
-  ];
-  const seen = new Set<string>();
-  const filters: Array<{ id: string; label: string }> = [];
-  for (const control of controls) {
-    const filter = getControlFilter(control, t);
-    if (!filter || seen.has(filter.id)) continue;
-    seen.add(filter.id);
-    filters.push(filter);
-  }
-  return filters;
-}
-
-function upsertLibraryFilterOption(options: Map<string, LibraryFilterOption>, id: string, label: string): void {
-  const existing = options.get(id);
-  if (existing) {
-    existing.count += 1;
-    return;
-  }
-  options.set(id, { id, label, count: 1 });
-}
-
-function mapLibraryFilterOptions(options: Map<string, LibraryFilterOption>): LibraryFilterOption[] {
-  return [...options.values()].sort((left, right) => left.label.localeCompare(right.label));
-}
-
-function gameHasLibraryActivity(game: GameInfo, playtimeData: PlaytimeData): boolean {
-  if (game.lastPlayed) return true;
-  const record = playtimeData[game.id];
-  if (!record) return false;
-  return Boolean(record.lastPlayedAt) || (record.totalSeconds ?? 0) > 0 || (record.sessionCount ?? 0) > 0;
-}
-
-function getLibraryFilterGroups(games: GameInfo[], playtimeData: PlaytimeData, t: LibraryTranslation): LibraryFilterGroup[] {
-  const playTypeOptions = new Map<string, LibraryFilterOption>();
-  const platformOptions = new Map<string, LibraryFilterOption>();
-  const controlOptions = new Map<string, LibraryFilterOption>();
-  const activityOptions = new Map<string, LibraryFilterOption>();
-
-  for (const game of games) {
-    const playTypeOption = getGamePlayTypeFilter(game, t);
-    if (playTypeOption) upsertLibraryFilterOption(playTypeOptions, playTypeOption.id, playTypeOption.label);
-
-    for (const option of getGameStoreFilters(game)) {
-      upsertLibraryFilterOption(platformOptions, option.id, option.label);
-    }
-
-    for (const option of getGameControlFilters(game, t)) {
-      upsertLibraryFilterOption(controlOptions, option.id, option.label);
-    }
-
-    const hasActivity = gameHasLibraryActivity(game, playtimeData);
-    upsertLibraryFilterOption(
-      activityOptions,
-      hasActivity ? "activity:played" : "activity:never-played",
-      hasActivity ? t("library.filterOptions.played") : t("library.filterOptions.neverPlayed"),
-    );
-  }
-
-  return [
-    { id: "play", label: t("library.filterGroups.playType"), options: mapLibraryFilterOptions(playTypeOptions) },
-    { id: "platform", label: t("library.filterGroups.platform"), options: mapLibraryFilterOptions(platformOptions) },
-    { id: "controls", label: t("library.filterGroups.controls"), options: mapLibraryFilterOptions(controlOptions) },
-    { id: "activity", label: t("library.filterGroups.activity"), options: mapLibraryFilterOptions(activityOptions) },
-  ].filter((group) => group.options.length > 0);
-}
-
-function getLibraryFilterGroupId(filterId: string): string {
-  const separatorIndex = filterId.indexOf(":");
-  return separatorIndex >= 0 ? filterId.slice(0, separatorIndex) : filterId;
-}
-
-function gameMatchesLibraryFilter(game: GameInfo, filterId: string, playtimeData: PlaytimeData, t: LibraryTranslation): boolean {
-  const [groupId, value] = filterId.split(":");
-  if (!value) return true;
-
-  if (groupId === "play") {
-    return getGamePlayTypeFilter(game, t)?.id === filterId;
-  }
-
-  if (groupId === "platform") {
-    return getGameStoreFilters(game).some((option) => option.id === filterId);
-  }
-
-  if (groupId === "controls") {
-    return getGameControlFilters(game, t).some((option) => option.id === filterId);
-  }
-
-  if (groupId === "activity") {
-    const hasActivity = gameHasLibraryActivity(game, playtimeData);
-    return value === "played" ? hasActivity : !hasActivity;
-  }
-
-  return true;
-}
-
-function gameMatchesLibraryFilters(game: GameInfo, selectedFilterIds: string[], playtimeData: PlaytimeData, t: LibraryTranslation): boolean {
-  if (selectedFilterIds.length === 0) return true;
-  const selectedByGroup = new Map<string, string[]>();
-  for (const filterId of selectedFilterIds) {
-    const groupId = getLibraryFilterGroupId(filterId);
-    selectedByGroup.set(groupId, [...(selectedByGroup.get(groupId) ?? []), filterId]);
-  }
-  for (const groupFilterIds of selectedByGroup.values()) {
-    if (!groupFilterIds.some((filterId) => gameMatchesLibraryFilter(game, filterId, playtimeData, t))) return false;
-  }
-  return true;
-}
-
-function getLibraryFilterOptionById(groups: LibraryFilterGroup[], id: string): LibraryFilterOption | undefined {
-  for (const group of groups) {
-    const option = group.options.find((candidate) => candidate.id === id);
-    if (option) return option;
-  }
-  return undefined;
-}
-
-function getControllerStoreFilterItems(games: GameInfo[], allStoresLabel: string): ControllerStoreFilterItem[] {
-  const stores = new Set<string>();
-  for (const game of games) {
-    for (const store of game.availableStores ?? []) {
-      if (store.trim()) stores.add(store);
-    }
-    for (const variant of game.variants) {
-      if (variant.store.trim()) stores.add(variant.store);
-    }
-  }
-
-  return [
-    { id: "library", title: allStoresLabel },
-    ...[...stores].sort((left, right) => left.localeCompare(right)).map((store) => ({ id: `store:${store}`, title: store })),
-  ];
-}
-
-function ControllerGameCard({
-  game,
-  isSelected,
-  selectedVariantId,
-  onSelect,
-  onPlay,
-}: {
-  game: GameInfo;
-  isSelected: boolean;
-  selectedVariantId?: string;
-  onSelect: () => void;
-  onPlay: () => void;
-}): JSX.Element {
-  const selectedVariant = game.variants.find((variant) => variant.id === selectedVariantId) ?? game.variants[game.selectedVariantIndex] ?? game.variants[0];
-  const store = selectedVariant?.store ?? game.availableStores?.[0] ?? "";
-  const StoreIcon = getStoreIconComponent(store);
-  const logoUrl = getGameLogoUrl(game);
-
-  return (
-    <m.button
-      type="button"
-      className={`controller-native-card${isSelected ? " selected" : ""}`}
-      onClick={onSelect}
-      onDoubleClick={onPlay}
-      aria-label={game.title}
-      animate={{ y: isSelected ? -3 : 0, scale: isSelected ? 1.025 : 1 }}
-      whileHover={{ y: -3, scale: 1.02 }}
-      whileTap={{ scale: 0.985 }}
-      transition={panelSpring}
-    >
-      <span className="controller-native-card-art">
-        {game.imageUrl ? <img src={game.imageUrl} alt="" loading="lazy" /> : <span className="controller-native-card-placeholder">{game.title.slice(0, 1)}</span>}
-      </span>
-      {store && (
-        <span className="controller-native-card-store" title={getStoreDisplayName(store)}>
-          <StoreIcon />
-        </span>
-      )}
-      <span className="controller-native-card-title">
-        {logoUrl ? <img src={logoUrl} alt={game.title} loading="lazy" /> : game.title}
-      </span>
-    </m.button>
-  );
 }
 
 export const LibraryPage = memo(function LibraryPage({
@@ -834,253 +476,37 @@ export const LibraryPage = memo(function LibraryPage({
 
   if (controllerMode) {
     const featuredGame = controllerFeaturedGames[controllerHeroIndex] ?? selectedControllerGame;
-    const heroImageUrl = featuredGame ? getControllerHeroBackgroundCandidates(featuredGame)[0] : undefined;
-    const heroLogoUrl = featuredGame ? getControllerHeroLogoUrl(featuredGame) : undefined;
-    const heroSelectedVariantId = featuredGame ? selectedVariantByGameId[featuredGame.id] : undefined;
-    const heroStoreLabel = featuredGame ? getSelectedVariantStoreLabel(featuredGame, selectedVariantByGameId[featuredGame.id], t("library.storeNotListed")) : "";
-    const featuredGameHasActiveSession = featuredGame ? gameMatchesActiveSession(featuredGame, activeSessionAppIds) : false;
-    const heroShouldBuy = Boolean(featuredGame && !featuredGameHasActiveSession && !featuredGame.isInLibrary);
-    const dotCount = Math.min(Math.max(controllerFeaturedGames.length, 1), 6);
-    const activeDotIndex = dotCount > 0 && controllerFeaturedGames.length > 0 ? Math.min(controllerHeroIndex, dotCount - 1) : 0;
-
     return (
-      <div className="library-page controller-library-page">
-        {isLoading ? (
-          <div className="library-empty-state controller-library-empty">
-            <Loader2 className="library-spinner" size={54} />
-            <p>{t("library.empty.loadingLibrary")}</p>
-          </div>
-        ) : libraryCount === 0 ? (
-          <div className="library-empty-state controller-library-empty">
-            <Gamepad2 className="library-empty-icon" size={64} />
-            <h3>{t("library.empty.libraryEmpty")}</h3>
-            <p>{t("library.empty.ownedGamesAppearHere")}</p>
-          </div>
-        ) : featuredGame ? (
-          <>
-            <section className="controller-hero" aria-label={featuredGame.title}>
-              <AnimatePresence initial={false} mode="popLayout">
-                {heroImageUrl ? (
-                  <m.img
-                    key={heroImageUrl}
-                    src={heroImageUrl}
-                    alt=""
-                    className="controller-hero-image"
-                    initial={{ opacity: 0, scale: 1.035 }}
-                    animate={{ opacity: 1, scale: 1 }}
-                    exit={{ opacity: 0, scale: 1.015 }}
-                    transition={pageTransition}
-                  />
-                ) : (
-                  <m.div
-                    key="controller-library-hero-placeholder"
-                    className="controller-hero-placeholder"
-                    initial={{ opacity: 0 }}
-                    animate={{ opacity: 1 }}
-                    exit={{ opacity: 0 }}
-                    transition={pageTransition}
-                  />
-                )}
-              </AnimatePresence>
-              <div className="controller-hero-scrim" />
-              <AnimatePresence initial={false} mode="wait">
-                <m.div
-                  key={featuredGame.id}
-                  className="controller-hero-content"
-                  initial={{ opacity: 0, y: 16 }}
-                  animate={{ opacity: 1, y: 0 }}
-                  exit={{ opacity: 0, y: -8 }}
-                  transition={pageTransition}
-                >
-                  {heroLogoUrl ? (
-                    <img src={heroLogoUrl} alt={featuredGame.title} className="controller-hero-logo" />
-                  ) : (
-                    <h1>{featuredGame.title}</h1>
-                  )}
-                  <div className="controller-hero-actions">
-                    <button
-                      type="button"
-                      className="controller-primary-action"
-                      onClick={() => {
-                        if (heroShouldBuy) {
-                          onBuyGame?.(featuredGame, heroSelectedVariantId);
-                          return;
-                        }
-                        onPlayGame(featuredGame);
-                      }}
-                    >
-                      {featuredGameHasActiveSession ? t("app.actions.resume") : heroShouldBuy ? t("app.actions.buy") : t("app.actions.play")}
-                    </button>
-                    {heroStoreLabel && <span className="controller-hero-variant-pill">{heroStoreLabel}</span>}
-                    <button type="button" className="controller-icon-action" aria-label={t("library.moreOptions")} onClick={() => cycleGameVariant(featuredGame)}>
-                      <MoreHorizontal size={30} />
-                    </button>
-                  </div>
-                </m.div>
-              </AnimatePresence>
-            </section>
-
-            <div className="controller-hero-dots" aria-hidden="true">
-              {Array.from({ length: dotCount }).map((_, index) => (
-                <span key={index} className={index === activeDotIndex ? "active" : ""} />
-              ))}
-            </div>
-
-            <section className="controller-library-strip" aria-label={t("library.title")}> 
-              <div className="controller-library-heading">
-                <h2>{t("library.controllerTitle")}</h2>
-                <span>{t("library.gameCount", { count: controllerGames.length })}</span>
-              </div>
-              {controllerGames.length === 0 ? (
-                <div className="library-empty-state controller-library-empty controller-library-empty--compact">
-                  <Search className="library-empty-icon" size={44} />
-                  <h3>{t("library.empty.noGamesFound")}</h3>
-                  <p>{t("library.empty.noGamesMatch", { query: searchQuery })}</p>
-                </div>
-              ) : (
-                <div className="controller-game-row" ref={controllerGameRowRef}>
-                  {controllerGames.map((game) => (
-                    <div key={game.id} className="controller-library-card" data-controller-game-id={game.id}>
-                      <ControllerGameCard
-                        game={game}
-                        isSelected={game.id === selectedGameId}
-                        onSelect={() => onSelectGame(game.id)}
-                        onPlay={() => onPlayGame(game)}
-                        selectedVariantId={selectedVariantByGameId[game.id]}
-                      />
-                    </div>
-                  ))}
-                </div>
-              )}
-            </section>
-
-            <div className="controller-bottom-hints" aria-hidden="true">
-              <div className="controller-hint"><span className="controller-button controller-button--a">A</span><span>{t("app.actions.select")}</span></div>
-              <div className="controller-hint"><span className="controller-button controller-button--b">B</span><span>{t("app.actions.back")}</span></div>
-              <div className="controller-hint"><span className="controller-button controller-button--y">Y</span><span>{t("library.filter")}</span></div>
-              <div className="controller-hint"><span className="controller-button controller-button--x">X</span><span>{t("app.actions.search")}</span></div>
-              <div className="controller-hint controller-hint--more"><span className="controller-menu-button"><Menu size={22} /></span><span>{t("library.moreOptions")}</span></div>
-            </div>
-
-            <AnimatePresence initial={false}>
-              {controllerStoreFilterOpen && (
-                <m.div
-                  className="controller-store-filter-overlay"
-                  role="dialog"
-                  aria-modal="true"
-                  aria-label={t("library.chooseStore")}
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  exit={{ opacity: 0 }}
-                  transition={pageTransition}
-                >
-                  <m.div
-                    className="controller-store-filter-panel"
-                    initial={{ opacity: 0, y: 16, scale: 0.985 }}
-                    animate={{ opacity: 1, y: 0, scale: 1 }}
-                    exit={{ opacity: 0, y: 10, scale: 0.99 }}
-                    transition={panelSpring}
-                  >
-                  <span className="controller-store-filter-eyebrow">{t("library.storeFilter")}</span>
-                  <h3>{t("library.chooseStore")}</h3>
-                  <p>{t("library.storeFilterHint")}</p>
-                  <div className="controller-store-filter-options">
-                    {controllerStoreFilterItems.map((item, index) => (
-                      <button
-                        key={item.id}
-                        type="button"
-                        className={`controller-store-filter-option${index === focusedControllerStoreFilterIndex ? " focused" : ""}`}
-                        onClick={() => {
-                          setFocusedControllerStoreFilterIndex(index);
-                          setControllerStoreFilterId(item.id);
-                          setControllerStoreFilterOpen(false);
-                        }}
-                      >
-                        {item.title}
-                      </button>
-                    ))}
-                  </div>
-                  </m.div>
-                </m.div>
-              )}
-            </AnimatePresence>
-
-            <AnimatePresence initial={false}>
-              {controllerSearchOpen && (
-                <m.div
-                  className="controller-search-overlay"
-                  role="dialog"
-                  aria-modal="true"
-                  aria-label={t("app.actions.search")}
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  exit={{ opacity: 0 }}
-                  transition={pageTransition}
-                >
-                  <m.div
-                    className="controller-search-panel"
-                    initial={{ opacity: 0, y: 16, scale: 0.985 }}
-                    animate={{ opacity: 1, y: 0, scale: 1 }}
-                    exit={{ opacity: 0, y: 10, scale: 0.99 }}
-                    transition={panelSpring}
-                  >
-                  <span className="controller-search-eyebrow">{t("app.actions.search")}</span>
-                  <input
-                    ref={controllerSearchInputRef}
-                    type="text"
-                    value={searchQuery}
-                    onChange={(event) => onSearchChange(event.target.value)}
-                    placeholder={t("library.searchPlaceholder")}
-                    className="controller-search-input"
-                  />
-                  <p>{t("app.actions.back")}</p>
-                  </m.div>
-                </m.div>
-              )}
-            </AnimatePresence>
-
-            <AnimatePresence initial={false}>
-              {detailsGame && (
-                <m.div
-                  className="controller-details-overlay"
-                  role="dialog"
-                  aria-modal="true"
-                  aria-label={detailsGame.title}
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  exit={{ opacity: 0 }}
-                  transition={pageTransition}
-                >
-                  <m.div
-                    className="controller-details-panel"
-                    initial={{ opacity: 0, y: 18, scale: 0.985 }}
-                    animate={{ opacity: 1, y: 0, scale: 1 }}
-                    exit={{ opacity: 0, y: 10, scale: 0.99 }}
-                    transition={panelSpring}
-                  >
-                  <h3>{detailsGame.title}</h3>
-                  <p className="controller-details-store">{t("library.selectedStore", { store: getGameStoreSummary(detailsGame, t("library.storeNotListed")) })}</p>
-                  <p className="controller-details-body">{detailsGame.description || detailsGame.longDescription || detailsGame.featureLabels?.join(" / ") || t("library.loadingGameDetails")}</p>
-                  <div className="controller-details-meta">
-                    {detailsGame.developerName && <span>{t("library.developer", { developer: detailsGame.developerName })}</span>}
-                    {detailsGame.publisherName && <span>{t("library.publisher", { publisher: detailsGame.publisherName })}</span>}
-                    {getPlayerSummary(detailsGame) && <span>{t("library.players", { players: getPlayerSummary(detailsGame) })}</span>}
-                    {detailsGame.supportedControls?.length ? <span>{t("library.controls", { controls: detailsGame.supportedControls.slice(0, 4).join(", ") })}</span> : null}
-                    {detailsGame.nvidiaTech?.length ? <span>{t("library.nvidiaTech", { tech: detailsGame.nvidiaTech.slice(0, 4).join(", ") })}</span> : null}
-                    {detailsGame.genres?.length ? <span>{t("library.genres", { genres: detailsGame.genres.slice(0, 4).join(", ") })}</span> : null}
-                    {detailsGame.contentRatings?.length ? <span>{t("library.rating", { rating: detailsGame.contentRatings.slice(0, 2).join(", ") })}</span> : null}
-                  </div>
-                  <div className="controller-details-actions">
-                    <button type="button" className="controller-primary-action" onClick={() => onPlayGame(detailsGame)}>{t("app.actions.play")}</button>
-                    <button type="button" className="controller-secondary-action" onClick={() => setDetailsGame(null)}>{t("app.actions.back")}</button>
-                  </div>
-                  </m.div>
-                </m.div>
-              )}
-            </AnimatePresence>
-          </>
-        ) : null}
-      </div>
+      <LibraryControllerView
+        isLoading={isLoading}
+        libraryCount={libraryCount}
+        searchQuery={searchQuery}
+        onSearchChange={onSearchChange}
+        selectedGameId={selectedGameId}
+        onSelectGame={onSelectGame}
+        onPlayGame={onPlayGame}
+        onBuyGame={onBuyGame}
+        selectedVariantByGameId={selectedVariantByGameId}
+        activeSessionAppIds={activeSessionAppIds}
+        featuredGame={featuredGame}
+        controllerFeaturedGames={controllerFeaturedGames}
+        controllerHeroIndex={controllerHeroIndex}
+        controllerGames={controllerGames}
+        controllerGameRowRef={controllerGameRowRef}
+        controllerStoreFilterOpen={controllerStoreFilterOpen}
+        controllerStoreFilterItems={controllerStoreFilterItems}
+        focusedControllerStoreFilterIndex={focusedControllerStoreFilterIndex}
+        onFocusControllerStoreFilter={setFocusedControllerStoreFilterIndex}
+        onSelectControllerStoreFilter={(itemId) => {
+          setControllerStoreFilterId(itemId);
+          setControllerStoreFilterOpen(false);
+        }}
+        controllerSearchOpen={controllerSearchOpen}
+        controllerSearchInputRef={controllerSearchInputRef}
+        detailsGame={detailsGame}
+        onCloseDetails={() => setDetailsGame(null)}
+        onCycleGameVariant={cycleGameVariant}
+      />
     );
   }
